+++ modules/AWSPowerShell/Cmdlets/MediaConnect/AWS.Tools.MediaConnect.Completers.psm1
@@ -104,6 +104,16 @@ $EMCN_Completers = {
             break
         }
 
+        # Amazon.MediaConnect.MaintenanceDay
+        {
+            ($_ -eq "New-EMCNFlow/Maintenance_MaintenanceDay") -Or
+            ($_ -eq "Update-EMCNFlow/Maintenance_MaintenanceDay")
+        }
+        {
+            $v = "Friday","Monday","Saturday","Sunday","Thursday","Tuesday","Wednesday"
+            break
+        }
+
         # Amazon.MediaConnect.MediaStreamType
         "Update-EMCNFlowMediaStream/MediaStreamType"
         {
@@ -166,6 +176,7 @@ $EMCN_map = @{
     "Attributes_Fmtp_ScanMode"=@("Update-EMCNFlowMediaStream")
     "Attributes_Fmtp_Tcs"=@("Update-EMCNFlowMediaStream")
     "EntitlementStatus"=@("Update-EMCNFlowEntitlement")
+    "Maintenance_MaintenanceDay"=@("New-EMCNFlow","Update-EMCNFlow")
     "MediaStreamType"=@("Update-EMCNFlowMediaStream")
     "Protocol"=@("Update-EMCNFlowOutput","Update-EMCNFlowSource")
     "SourceFailoverConfig_FailoverMode"=@("New-EMCNFlow","Update-EMCNFlow")