+++ modules/AWSPowerShell/Cmdlets/QuickSight/Basic/Find-QSGroup-Cmdlet.cs
@@ -28,16 +28,17 @@ using Amazon.QuickSight.Model;
 namespace Amazon.PowerShell.Cmdlets.QS
 {
     /// <summary>
-    /// Lists all user groups in Amazon QuickSight.<br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
+    /// Use the <code>SearchGroups</code> operation to search groups in a specified Amazon
+    /// QuickSight namespace using the supplied filters.<br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
     /// </summary>
-    [Cmdlet("Get", "QSGroupList")]
+    [Cmdlet("Find", "QSGroup")]
     [OutputType("Amazon.QuickSight.Model.Group")]
-    [AWSCmdlet("Calls the Amazon QuickSight ListGroups API operation.", Operation = new[] {"ListGroups"}, SelectReturnType = typeof(Amazon.QuickSight.Model.ListGroupsResponse))]
-    [AWSCmdletOutput("Amazon.QuickSight.Model.Group or Amazon.QuickSight.Model.ListGroupsResponse",
+    [AWSCmdlet("Calls the Amazon QuickSight SearchGroups API operation.", Operation = new[] {"SearchGroups"}, SelectReturnType = typeof(Amazon.QuickSight.Model.SearchGroupsResponse))]
+    [AWSCmdletOutput("Amazon.QuickSight.Model.Group or Amazon.QuickSight.Model.SearchGroupsResponse",
         "This cmdlet returns a collection of Amazon.QuickSight.Model.Group objects.",
-        "The service call response (type Amazon.QuickSight.Model.ListGroupsResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+        "The service call response (type Amazon.QuickSight.Model.SearchGroupsResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
     )]
-    public partial class GetQSGroupListCmdlet : AmazonQuickSightClientCmdlet, IExecutor
+    public partial class FindQSGroupCmdlet : AmazonQuickSightClientCmdlet, IExecutor
     {
         
         #region Parameter AwsAccountId
@@ -58,10 +59,28 @@ namespace Amazon.PowerShell.Cmdlets.QS
         public System.String AwsAccountId { get; set; }
         #endregion
         
+        #region Parameter Filter
+        /// <summary>
+        /// <para>
+        /// <para>The structure for the search filters that you want to apply to your search.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyCollection]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        [Alias("Filters")]
+        public Amazon.QuickSight.Model.GroupSearchFilter[] Filter { get; set; }
+        #endregion
+        
         #region Parameter Namespace
         /// <summary>
         /// <para>
-        /// <para>The namespace that you want a list of groups from.</para>
+        /// <para>The namespace that you want to search.</para>
         /// </para>
         /// </summary>
         #if !MODULAR
@@ -78,18 +97,12 @@ namespace Amazon.PowerShell.Cmdlets.QS
         #region Parameter MaxResult
         /// <summary>
         /// <para>
-        /// <para>The maximum number of results to return.</para>
+        /// <para>The maximum number of results to return from this request.</para>
         /// </para>
-        /// <para>
-        /// <br/><b>Note:</b> In AWSPowerShell and AWSPowerShell.NetCore this parameter is used to limit the total number of items returned by the cmdlet.
-        /// <br/>In AWS.Tools this parameter is simply passed to the service to specify how many items should be returned by each service call.
-        /// <br/>Pipe the output of this cmdlet into Select-Object -First to terminate retrieving data pages early and control the number of items returned.
-        /// </para>
-        /// <para>If a value for this parameter is not specified the cmdlet will use a default value of '<b>100</b>'.</para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
-        [Alias("MaxItems","MaxResults")]
-        public int? MaxResult { get; set; }
+        [Alias("MaxResults")]
+        public System.Int32? MaxResult { get; set; }
         #endregion
         
         #region Parameter NextToken
@@ -109,8 +122,8 @@ namespace Amazon.PowerShell.Cmdlets.QS
         #region Parameter Select
         /// <summary>
         /// Use the -Select parameter to control the cmdlet output. The default value is 'GroupList'.
-        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.QuickSight.Model.ListGroupsResponse).
-        /// Specifying the name of a property of type Amazon.QuickSight.Model.ListGroupsResponse will result in that property being returned.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.QuickSight.Model.SearchGroupsResponse).
+        /// Specifying the name of a property of type Amazon.QuickSight.Model.SearchGroupsResponse will result in that property being returned.
         /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -149,7 +162,7 @@ namespace Amazon.PowerShell.Cmdlets.QS
             #pragma warning disable CS0618, CS0612 //A class member was marked with the Obsolete attribute
             if (ParameterWasBound(nameof(this.Select)))
             {
-                context.Select = CreateSelectDelegate<Amazon.QuickSight.Model.ListGroupsResponse, GetQSGroupListCmdlet>(Select) ??
+                context.Select = CreateSelectDelegate<Amazon.QuickSight.Model.SearchGroupsResponse, FindQSGroupCmdlet>(Select) ??
                     throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
                 if (this.PassThru.IsPresent)
                 {
@@ -168,23 +181,17 @@ namespace Amazon.PowerShell.Cmdlets.QS
                 WriteWarning("You are passing $null as a value for parameter AwsAccountId which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
-            context.MaxResult = this.MaxResult;
+            if (this.Filter != null)
+            {
+                context.Filter = new List<Amazon.QuickSight.Model.GroupSearchFilter>(this.Filter);
+            }
             #if MODULAR
-            if (!ParameterWasBound(nameof(this.MaxResult)))
+            if (this.Filter == null && ParameterWasBound(nameof(this.Filter)))
             {
-                WriteVerbose("MaxResult parameter unset, using default value of '100'");
-                context.MaxResult = 100;
+                WriteWarning("You are passing $null as a value for parameter Filter which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
-            #if !MODULAR
-            if (ParameterWasBound(nameof(this.MaxResult)) && this.MaxResult.HasValue)
-            {
-                WriteWarning("AWSPowerShell and AWSPowerShell.NetCore use the MaxResult parameter to limit the total number of items returned by the cmdlet." +
-                    " This behavior is obsolete and will be removed in a future version of these modules. Pipe the output of this cmdlet into Select-Object -First to terminate" +
-                    " retrieving data pages early and control the number of items returned. AWS.Tools already implements the new behavior of simply passing MaxResult" +
-                    " to the service to specify how many items should be returned by each service call.");
-            }
-            #endif
+            context.MaxResult = this.MaxResult;
             context.Namespace = this.Namespace;
             #if MODULAR
             if (this.Namespace == null && ParameterWasBound(nameof(this.Namespace)))
@@ -203,7 +210,6 @@ namespace Amazon.PowerShell.Cmdlets.QS
         
         #region IExecutor Members
         
-        #if MODULAR
         public object Execute(ExecutorContext context)
         {
             var cmdletContext = context as CmdletContext;
@@ -212,15 +218,19 @@ namespace Amazon.PowerShell.Cmdlets.QS
             #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
             
             // create request and set iteration invariants
-            var request = new Amazon.QuickSight.Model.ListGroupsRequest();
+            var request = new Amazon.QuickSight.Model.SearchGroupsRequest();
             
             if (cmdletContext.AwsAccountId != null)
             {
                 request.AwsAccountId = cmdletContext.AwsAccountId;
             }
+            if (cmdletContext.Filter != null)
+            {
+                request.Filters = cmdletContext.Filter;
+            }
             if (cmdletContext.MaxResult != null)
             {
-                request.MaxResults = AutoIterationHelpers.ConvertEmitLimitToServiceTypeInt32(cmdletContext.MaxResult.Value);
+                request.MaxResults = cmdletContext.MaxResult.Value;
             }
             if (cmdletContext.Namespace != null)
             {
@@ -273,107 +283,6 @@ namespace Amazon.PowerShell.Cmdlets.QS
             
             return null;
         }
-        #else
-        public object Execute(ExecutorContext context)
-        {
-            var cmdletContext = context as CmdletContext;
-            var useParameterSelect = this.Select.StartsWith("^") || this.PassThru.IsPresent;
-            
-            // create request and set iteration invariants
-            var request = new Amazon.QuickSight.Model.ListGroupsRequest();
-            if (cmdletContext.AwsAccountId != null)
-            {
-                request.AwsAccountId = cmdletContext.AwsAccountId;
-            }
-            if (cmdletContext.Namespace != null)
-            {
-                request.Namespace = cmdletContext.Namespace;
-            }
-            
-            // Initialize loop variants and commence piping
-            System.String _nextToken = null;
-            int? _emitLimit = null;
-            int _retrievedSoFar = 0;
-            if (AutoIterationHelpers.HasValue(cmdletContext.NextToken))
-            {
-                _nextToken = cmdletContext.NextToken;
-            }
-            if (cmdletContext.MaxResult.HasValue)
-            {
-                // The service has a maximum page size of 100. If the user has
-                // asked for more items than page max, and there is no page size
-                // configured, we rely on the service ignoring the set maximum
-                // and giving us 100 items back. If a page size is set, that will
-                // be used to configure the pagination.
-                // We'll make further calls to satisfy the user's request.
-                _emitLimit = cmdletContext.MaxResult;
-            }
-            var _userControllingPaging = this.NoAutoIteration.IsPresent || ParameterWasBound(nameof(this.NextToken));
-            
-            var client = Client ?? CreateClient(_CurrentCredentials, _RegionEndpoint);
-            do
-            {
-                request.NextToken = _nextToken;
-                if (_emitLimit.HasValue)
-                {
-                    int correctPageSize = Math.Min(100, _emitLimit.Value);
-                    request.MaxResults = AutoIterationHelpers.ConvertEmitLimitToInt32(correctPageSize);
-                }
-                else if (!ParameterWasBound(nameof(this.MaxResult)))
-                {
-                    request.MaxResults = AutoIterationHelpers.ConvertEmitLimitToInt32(100);
-                }
-                
-                CmdletOutput output;
-                
-                try
-                {
-                    
-                    var response = CallAWSServiceOperation(client, request);
-                    object pipelineOutput = null;
-                    if (!useParameterSelect)
-                    {
-                        pipelineOutput = cmdletContext.Select(response, this);
-                    }
-                    output = new CmdletOutput
-                    {
-                        PipelineOutput = pipelineOutput,
-                        ServiceResponse = response
-                    };
-                    int _receivedThisCall = response.GroupList.Count;
-                    
-                    _nextToken = response.NextToken;
-                    _retrievedSoFar += _receivedThisCall;
-                    if (_emitLimit.HasValue)
-                    {
-                        _emitLimit -= _receivedThisCall;
-                    }
-                }
-                catch (Exception e)
-                {
-                    if (_retrievedSoFar == 0 || !_emitLimit.HasValue)
-                    {
-                        output = new CmdletOutput { ErrorResponse = e };
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-                
-                ProcessOutput(output);
-            } while (!_userControllingPaging && AutoIterationHelpers.HasValue(_nextToken) && (!_emitLimit.HasValue || _emitLimit.Value >= 1));
-            
-            
-            if (useParameterSelect)
-            {
-                WriteObject(cmdletContext.Select(null, this));
-            }
-            
-            
-            return null;
-        }
-        #endif
         
         public ExecutorContext CreateContext()
         {
@@ -384,15 +293,15 @@ namespace Amazon.PowerShell.Cmdlets.QS
         
         #region AWS Service Operation Call
         
-        private Amazon.QuickSight.Model.ListGroupsResponse CallAWSServiceOperation(IAmazonQuickSight client, Amazon.QuickSight.Model.ListGroupsRequest request)
+        private Amazon.QuickSight.Model.SearchGroupsResponse CallAWSServiceOperation(IAmazonQuickSight client, Amazon.QuickSight.Model.SearchGroupsRequest request)
         {
-            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon QuickSight", "ListGroups");
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon QuickSight", "SearchGroups");
             try
             {
                 #if DESKTOP
-                return client.ListGroups(request);
+                return client.SearchGroups(request);
                 #elif CORECLR
-                return client.ListGroupsAsync(request).GetAwaiter().GetResult();
+                return client.SearchGroupsAsync(request).GetAwaiter().GetResult();
                 #else
                         #error "Unknown build edition"
                 #endif
@@ -413,10 +322,11 @@ namespace Amazon.PowerShell.Cmdlets.QS
         internal partial class CmdletContext : ExecutorContext
         {
             public System.String AwsAccountId { get; set; }
-            public int? MaxResult { get; set; }
+            public List<Amazon.QuickSight.Model.GroupSearchFilter> Filter { get; set; }
+            public System.Int32? MaxResult { get; set; }
             public System.String Namespace { get; set; }
             public System.String NextToken { get; set; }
-            public System.Func<Amazon.QuickSight.Model.ListGroupsResponse, GetQSGroupListCmdlet, object> Select { get; set; } =
+            public System.Func<Amazon.QuickSight.Model.SearchGroupsResponse, FindQSGroupCmdlet, object> Select { get; set; } =
                 (response, cmdlet) => response.GroupList;
         }