+++ modules/AWSPowerShell/Cmdlets/QuickSight/Basic/New-QSGroup-Cmdlet.cs
@@ -32,7 +32,7 @@ namespace Amazon.PowerShell.Cmdlets.QS
     /// 
     ///  
     /// <para>
-    /// The permissions resource is <code>arn:aws:quicksight:us-east-1:<i>&lt;relevant-aws-account-id&gt;</i>:group/default/<i>&lt;group-name&gt;</i></code>.
+    /// The permissions resource is <code>arn:aws:quicksight:&lt;your-region&gt;:<i>&lt;relevant-aws-account-id&gt;</i>:group/default/<i>&lt;group-name&gt;</i></code>.
     /// </para><para>
     /// The response is a group object.
     /// </para>
@@ -95,7 +95,7 @@ namespace Amazon.PowerShell.Cmdlets.QS
         #region Parameter Namespace
         /// <summary>
         /// <para>
-        /// <para>The namespace. Currently, you should set this to <code>default</code>.</para>
+        /// <para>The namespace that you want the group to be a part of.</para>
         /// </para>
         /// </summary>
         #if !MODULAR