+++ modules/AWSPowerShell/Cmdlets/QuickSight/Basic/Remove-QSGroupMembership-Cmdlet.cs
@@ -94,7 +94,7 @@ namespace Amazon.PowerShell.Cmdlets.QS
         #region Parameter Namespace
         /// <summary>
         /// <para>
-        /// <para>The namespace. Currently, you should set this to <code>default</code>.</para>
+        /// <para>The namespace of the group that you want to remove a user from.</para>
         /// </para>
         /// </summary>
         #if !MODULAR