+++ modules/AWSPowerShell/Cmdlets/RAM/Basic/Get-RAMResourceShare-Cmdlet.cs
@@ -67,7 +67,9 @@ namespace Amazon.PowerShell.Cmdlets.RAM
         /// <summary>
         /// <para>
         /// <para>Specifies that you want to retrieve details of only those resource shares that match
-        /// the following:</para><ul><li><para><b><code>SELF</code></b> – resources that you are sharing</para></li><li><para><b><code>OTHER-ACCOUNTS</code></b> – resources that other accounts share with you</para></li></ul>
+        /// the following:</para><ul><li><para><b><code>SELF</code></b> – resource shares that your account shares with other
+        /// accounts</para></li><li><para><b><code>OTHER-ACCOUNTS</code></b> – resource shares that other accounts share
+        /// with your account</para></li></ul>
         /// </para>
         /// </summary>
         #if !MODULAR